Added configuration of floating-point instantiation to configure.ac . Changed Real typedefs to Float. Renamed MPFloat to FloatMP. Changed test_arithmetic and test_function to test_rational and test_float. Split export_numeric to export_integer, export_rational and export_float.

+++ python/test/test_numeric.py
@@ -22,7 +22,7 @@
 
 
 from ariadne import *
-x=Float64(2.25)
+x=Float(2.25)
     
 n=2
 d=2.125
